Simplify example 52 to showcase using a custom OpenAIClient in general instead of using one specifically for APIM
